+++ source-code/core/src/lint/rules/brandingRule.ts
@@ -1,12 +1,12 @@
-import type { Reporter } from '../reporter.js'
+import type { Context } from '../context.js'
 import { createRule } from '../rule.js'
 
 export const createBrandingRule = (brandName: string, incorrectBrandingNames: string[]) => createRule('inlang.brandingRule', () => {
-	let reporter: Reporter
+	let context: Context
 
 	return {
 		initialize: (config) => {
-			reporter = config.reporter
+			context = config.context
 		},
 		visitors: {
 			Pattern: ({ target }) => {
@@ -16,7 +16,7 @@ export const createBrandingRule = (brandName: string, incorrectBrandingNames: st
 					.flatMap(element => incorrectBrandingNames.filter(word => element.value.includes(word)))
 
 				for (const incorrectlyBrandedElement of incorrectlyBrandedWords) {
-					reporter.report(target, `Element '${incorrectlyBrandedElement}' is incorrectly branded and should be replaced with '${brandName}'`)
+					context.report(target, `Element '${incorrectlyBrandedElement}' is incorrectly branded and should be replaced with '${brandName}'`)
 				}
 			}
 		},
